move all traits into a separate module
